fix cache statistics test to implement jsr 107 spec cache statistics table.
Added test that expire on creation does not result is a cache statistic put.  (as documented in getExpiryForCreation returing Duration#ZERO should not be put in cache).

Implemented iteration.

Ensured that all average ms checks are only checking for 0.0 or greater.  Reduces testing of accessing average cache statistics timing to just not throwing an exception.